requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
waitress==3.0.0
openai
